+++ concourse-plugin-core/src/main/java/com/cinchapi/concourse/server/plugin/data/LazyTrackingTObjectResultDataset.java
@@ -159,6 +159,11 @@ public class LazyTrackingTObjectResultDataset extends TObjectResultDataset {
         return data.toString();
     }
 
+    @Override
+    protected Map<TObject, Set<Long>> invertNullSafe(String attribute) {
+        return tracking().invertNullSafe(attribute);
+    }
+
     /**
      * Populate {@link #tracking}.
      */