+++ concourse-plugin-core/src/test/java/com/cinchapi/concourse/server/plugin/data/LazyTrackingTObjectResultDatasetTest.java
@@ -19,6 +19,7 @@ import java.nio.ByteBuffer;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map.Entry;
+import java.util.stream.Collectors;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -124,13 +125,35 @@ public class LazyTrackingTObjectResultDatasetTest
                 }
             }
         }
-
         PluginSerializer serializer = new PluginSerializer();
         ByteBuffer bytes = serializer.serialize(dataset);
         ComplexTObject complex = ComplexTObject.fromJavaObject(bytes);
         ByteBuffer bytes2 = complex.getJavaObject();
         Object obj = serializer.deserialize(bytes2);
         Assert.assertEquals(dataset, obj);
+    }
+
+    @Test
+    public void testInvertRepro() {
+        String json = FileOps.read(Resources.getAbsolutePath("/data.json"));
+        List<Multimap<String, Object>> objects = Convert.anyJsonToJava(json);
+        for (Multimap<String, Object> object : objects) {
+            long entity = Time.now();
+            dataset.put(entity,
+                    object.asMap().entrySet().stream()
+                            .collect(Collectors.toMap(Entry::getKey,
+                                    e -> e.getValue().stream()
+                                            .map(Convert::javaToThrift)
+                                            .collect(Collectors.toSet()))));
+        }
+        PluginSerializer serializer = new PluginSerializer();
+        ByteBuffer bytes = serializer.serialize(dataset);
+        ComplexTObject complex = ComplexTObject.fromJavaObject(bytes);
+        ByteBuffer bytes2 = complex.getJavaObject();
+        dataset = serializer.deserialize(bytes2);
+        ObjectResultDataset dataset2 = new ObjectResultDataset(dataset);
+        Assert.assertEquals(dataset.toString(), dataset2.toString());
+
     }
 
 }